Add strictArgs
Defines if there can be arbitrary variables. If set to true the command cannot have any unknown variables. The arguments will either be a subcommand or a flag, if not an error is thrown. This means that flags are ignored and can be used. 

It is suggested that this is set to true if you only want flags as args.